fix compilation due to macro usage
